backend/requirements.txt
# FastAPI Backend Requirements
# Texas Hold'em Poker API with LLM Integration

# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# WebSocket support
websockets==12.0

# HTTP client for external APIs (LLM integration)
httpx==0.25.2
requests==2.31.0

# Environment variables and configuration
python-dotenv==1.0.0

# JSON and data validation
python-multipart==0.0.6

# Async database support (for future use)
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0  # PostgreSQL async driver
aiosqlite==0.19.0  # SQLite async driver

# Date/time utilities
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # for testing FastAPI

# Existing poker system dependencies
# (These should already be installed in your main project)
texasholdem>=1.0.0
colorama>=0.4.6
openai>=1.0.0

# Production deployment
gunicorn==21.2.0

# Security and CORS
python-jose[cryptography]==3.3.0
bcrypt==4.1.1

# Optional: Redis for caching/session management
redis==5.0.1
aioredis==2.0.1

# Optional: Prometheus metrics
prometheus-client==0.19.0 